requirements.txt
# Security: Using exact version pins to prevent supply chain attacks
discord.py==2.4.0
beautifulsoup4==4.12.3
requests==2.32.4  # Fixed CVE-2024-47081 (.netrc credential leak)
deep-translator==1.11.4  # PYSEC-2022-252 (2022 supply chain attack - monitored)
python-dotenv==1.0.1
lxml==5.3.0
chardet==5.2.0
